rename files to still be next to each other in publiq-ui
